FEAT: Custom Error for External Program Failure
Parses error raised by Nushell during external program (Git) execution
to provide a cleaner front-end for possible function failure.
Nushell's internal error, provided in JSON format, is unpacked to
retrieve critical information such as the error and help messages.

Error is marked as raised by the function call through setting the
`label.span` of the error struct, which points to the latest "file";
when interactively running this function in REPL, the "latest file"
referred to in this error struct would be the function call,
which would dereference the "span" used in the error to information on
the function call.
This renders a cleaner error output that shows the function call rather
than the default (internal) error, which showed the exact line of error
within the function body.
Note that `label.span.end` is hard-coded to three characters after
`label.span.start`; since this is a custom function that poses as a Git
subcommand and depends solely on Git, we can safely say that any errors
from external commands are raised due to problems with Git, and as such
this error struct marks `git` and ONLY `git` as the source of error.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>